# src/swagger/SwaggerOutput.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Library API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Server"
    },
    {
      "url": "http://localhost:3000/sit",
      "description": "Sit Server"
    },
    {
      "url": "http://localhost:3000/prod",
      "description": "Prod Server"
    }
  ],
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "get all user",
        "parameters": [
          {
            "$ref": "#/parameters/search_val"
          },
          {
            "$ref": "#/parameters/sort_field"
          },
          {
            "$ref": "#/parameters/sort_order"
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/per_page"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "create user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "get user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "description": "update user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "delete user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/album": {
      "get": {
        "tags": [
          "Album"
        ],
        "description": "get all album",
        "parameters": [
          {
            "$ref": "#/parameters/search_val"
          },
          {
            "$ref": "#/parameters/sort_field"
          },
          {
            "$ref": "#/parameters/sort_order"
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/per_page"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Album"
        ],
        "description": "create album",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/album"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "fullname": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john_doe@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "johndoe123"
          },
          "role": {
            "type": "string",
            "example": "subscriber"
          },
          "album_id": {
            "type": "number",
            "example": 1
          },
          "general_info": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "name"
              },
              "lastname": {
                "type": "string",
                "example": "lastname"
              },
              "age": {
                "type": "number",
                "example": 20
              }
            }
          }
        }
      },
      "album": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "album_name"
          },
          "list_song": {
            "type": "array",
            "example": [
              "song1",
              "song2"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      },
      "petstore_auth": {
        "type": "oauth2",
        "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
        "flow": "implicit",
        "scopes": {
          "read_pets": "read your pets",
          "write_pets": "modify pets in your account"
        }
      }
    }
  },
  "parameters": {
    "search_val": {
      "name": "search_val",
      "type": "string",
      "in": "query",
      "description": "ค่าที่ต้องการ search"
    },
    "sort_field": {
      "name": "sort_field",
      "schema": {
        "type": "array",
        "items": {
          "type": "string",
          "required": true
        },
        "example": [
          "createdAt"
        ]
      },
      "in": "query",
      "description": "ชื่อ field ที่ต้องการ sort (รองรับทุก field)"
    },
    "sort_order": {
      "name": "sort_order",
      "schema": {
        "type": "array",
        "items": {
          "type": "integer",
          "required": true
        },
        "example": [
          "-1"
        ]
      },
      "in": "query",
      "description": "1 น้อยไปมาก, -1 มากไปน้อย (รองรับทุก field เป็น stable sort จำนวนที่ใส่ต้องตรงกับจำนวน sort_field)"
    },
    "page": {
      "name": "page",
      "type": "integer",
      "in": "query",
      "description": "pagination (page to show)",
      "required": true,
      "example": 1
    },
    "per_page": {
      "name": "per_page",
      "type": "integer",
      "in": "query",
      "description": "pagination (number of doc to show in each page)",
      "required": true,
      "example": 10
    }
  }
}